Run-with-Heroku/heroku_backend/requirements.txt
asgiref==3.5.0
boto3==1.22.4
botocore==1.25.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.1
jmespath==1.0.0
Markdown==3.3.6
Pillow==9.0.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyqt5-tools==5.15.4.3.2
PyQtChart==5.15.5
PyQtChart-Qt5==5.15.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.0.0
zipp==3.7.0
